requirements.txt
anthropic
python-dotenv
requests
fastapi
uvicorn[standard]
beautifulsoup4
